Add rename script and associated github action.
This commit adds a github action which will run a rename script when this
repository is used as a template. This will streamline the KBase module
creation process. Then both the script and the action deletes itself and
commits the changes. What is left is a module ready to be used as a starting
point for the KB SDK tutorial.